Mushroom classification 🍄
In this project, we will use a public dataset from kaggle.com, the “Mushroom Classification” dataset, and try to figure out if a mushroom is edible or not. This dataset includes descriptions of hypothetical samples corresponding to 23 species of gilled mushrooms. Each species is identified as edible or poisonous. When it comes to mushrooms, there is no simple rule for determining the edibility.
We will tackle this classification problem using logistic regression, ridge classifier, decision tree, Naive Bayes, and neural networks. After comparing the results of each model, we will find out the best performing one.